+++ scim2_tester/checkers/__init__.py
@@ -20,6 +20,9 @@ from .resource_get import object_query
 from .resource_get import object_query_without_id
 from .resource_post import object_creation
 from .resource_put import object_replacement
+from .resource_query_attributes import object_list_with_attributes
+from .resource_query_attributes import object_query_with_attributes
+from .resource_query_attributes import search_with_attributes
 from .resource_types import access_invalid_resource_type
 from .resource_types import query_all_resource_types
 from .resource_types import query_resource_type_by_id
@@ -49,6 +52,9 @@ __all__ = [
     "object_creation",
     "object_query",
     "object_query_without_id",
+    "object_query_with_attributes",
+    "object_list_with_attributes",
+    "search_with_attributes",
     "object_replacement",
     "object_deletion",
     "resource_type_tests",